Share graphql schema from server w/ client in k8s
- Previously, there was not a sharing of the server schema with the
client, meaning that any subsequent deploys that changed the graphql
schema would not appropriately be picked up by the client.  This commit
changes the approach in kubernetes and the client docker image by
leaving asset building to part of the kubernetes command step - there
are performance downsides to this, and the image is likely larger, but
no other good solutions presented themselves.

- Remove now no longer needed local dockerfile for client

- Add some volumes to be used for temporary build files storage in
between client stages; add init container for building client assets

- Generate schema on server in init container